Backend/ECommerce/requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
Django==5.0.6
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
pillow==10.3.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
